+++ __init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.0.0"
+__version__ = "1.0.1"
 
 from D4CMPP2.src.api.training import train as train
 from D4CMPP2.src.api.legacy_grid_search import grid_search as grid_search
@@ -13,4 +13,3 @@ from D4CMPP2.networks import (
     MolecularNetwork as MolecularNetwork,
     register_network as register_network,
 )
-